+++ multi-token-edit/scripts/presets/fauxSceneImporter.js
@@ -318,7 +318,7 @@ class FauxSceneImporter extends HandlebarsApplicationMixin(ApplicationV2) {
         if (type === 'create') {
             callback = (tags) => {
                 this.toCreateFauxScenes.forEach((entry) => {
-                    entry.tags = [...new Set([...tags, ...[entry.tags ?? []]])];
+                    entry.tags = [...new Set([...tags, ...(entry.tags ?? [])])];
                 });
             };
         } else if (type === 'rename' || type === 'dangling') {